:truck: 이종수: 유효성 관련 폴더(validation) 재구조화
