+++ client/src/interfaces/email/Email.interface.ts
