+++ client/src/interfaces/password/Password.interface.ts
