+++ client/src/interfaces/validation/language/kor/KoreanValidator.interface.ts
