+++ client/src/interfaces/validation/space/WhitespaceValidator.interface.ts
