+++ client/src/modules/validation/utils/ValidationUtils.ts
