+++ client/src/pages/admin/memberInfo.tsx
@@ -3,7 +3,7 @@
 import { GetServerSideProps } from "next";
 import AdminNav from "src/components/nav/admin/adminNav";
 import MemberInfoTable from "src/components/table/member/MemberInfoTable";
-import IMemberInfo from "../../interfaces/MemberInfo.interface";
+import IMemberInfo from "../../interfaces/member/MemberInfo.interface";
 
 interface Props {
   members: IMemberInfo[];
